update jsmn
Pull in all fixes in the last six years from upstream. Contains two
documentation tweaks and a fix for JSMN_PARENT_LINKS (a mode we are not using
yet). The only functional fix we skip is the patch that makes strict mode
stricter [143] because our use of jsmn is about non-strict "easy to write for
humans" JSON (i.e. unquoted keys).

Don't copy the single header edition that was released after v1.0.0 verbatim
because it is mostly about stylistic changes. The existing coding style better
aligns with our code. Just handpick actual fixes and functional changes.

[143]: https://github.com/zserge/jsmn/pull/143